# estudio_ia_videos/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Estudio IA Videos API
  description: |
    API for converting PPTX presentations to AI-generated videos with avatars,
    text-to-speech narration, background music, and automatic subtitles.

    ## Authentication
    All API endpoints require authentication using an API key.
    Include your API key in the request headers:
    - `Authorization: Bearer {your_api_key}`
    - OR `X-API-Key: {your_api_key}`

    ## Rate Limits
    - Free plan: 10 requests/minute, 100 requests/month
    - Starter: 30 requests/minute, 1,000 requests/month
    - Pro: 100 requests/minute, 10,000 requests/month
    - Enterprise: 1,000 requests/minute, 100,000 requests/month

    ## Webhooks
    For long-running operations like video generation, you can provide a
    webhook URL to receive status updates instead of polling.
  version: 1.0.0
  contact:
    name: API Support
    email: api@estudioiavideos.com
  license:
    name: Proprietary
servers:
  - url: https://app.estudioiavideos.com/api/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...
paths:
...
  /music:
    get:
      tags:
        - Music
      summary: List background music
      description: Returns available royalty-free background music tracks
      operationId: listMusic
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [corporate, training, motivational, ambient, technology, documentary, uplifting, calm]
        - name: mood
          in: query
          schema:
            type: string
            enum: [professional, energetic, calm, inspiring, serious, friendly, modern, classic, minimal, dramatic]
        - name: search
          in: query
          schema:
            type: string
        - name: min_duration
          in: query
          schema:
            type: integer
        - name: max_duration
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of music tracks
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  categories:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        count:
                          type: integer
                  tracks:
                    type: array
                    items:
                      $ref: '#/components/schemas/MusicTrack'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    MusicTrack:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        artist:
          type: string
        category:
          type: string
        duration:
          type: integer
          description: Duration in seconds
        duration_formatted:
          type: string
          example: "3:45"
        bpm:
          type: integer
        url:
          type: string
          format: uri
        preview_url:
          type: string
          format: uri
        license:
          type: string
        tags:
          type: array
          items:
            type: string
        is_loop:
          type: boolean
...